Check if inside tmux session before prompt

+++ ks
@@ -1,7 +1,11 @@
 #! /bin/bash
 
-echo -n "kill session (Y/n)? "
-read answer
-if [ "$answer" == "Y" ]; then
-  tmux kill-session
+if [[ $TMUX ]] ; then
+  echo -n "kill session (Y/n)? "
+  read answer
+  if [ "$answer" == "Y" ]; then
+    tmux kill-session
+  fi
+else
+  echo -e "Not in tmux session"
 fi
